requirements.txt
pycrptodome
pillow
bcrypt
PyJWT
numpy
pandas
PyYAML
fastapi
pydantic
uvicorn
aiohttp
python-dotenv